[channels,rdp2tcp] move to custom channels
* move rdp2tcp to custom channels (loadable at runtime)
* remove code in core library (use /svc:rdp2tcp to load)
