--- channels/rdp2tcp/ChannelOptions.cmake
@@ -1,10 +1,0 @@
-set(OPTION_DEFAULT OFF)
-set(OPTION_CLIENT_DEFAULT ON)
-set(OPTION_SERVER_DEFAULT OFF)
-
-define_channel_options(NAME "rdp2tcp" TYPE "static"
-	DESCRIPTION "Tunneling TCP over RDP"
-	DEFAULT ${OPTION_DEFAULT})
-
-define_channel_client_options(${OPTION_CLIENT_DEFAULT})
-define_channel_server_options(${OPTION_SERVER_DEFAULT})
