+++ custom-channels/CMakeLists.txt
@@ -1,2 +1,3 @@
-add_subdirectory(sshagent)
 add_subdirectory(ainput)
+add_subdirectory(rdp2tcp)
+add_subdirectory(sshagent)
