+++ include/freerdp/settings.h
@@ -936,7 +936,6 @@ extern "C"
 #define FreeRDP_SupportDisplayControl (5185)
 #define FreeRDP_SupportGeometryTracking (5186)
 #define FreeRDP_SupportVideoOptimized (5188)
-#define FreeRDP_RDP2TCPArgs (5189)
 #define FreeRDP_TcpKeepAlive (5190)
 #define FreeRDP_TcpKeepAliveRetries (5191)
 #define FreeRDP_TcpKeepAliveDelay (5192)
@@ -1682,9 +1681,9 @@ extern "C"
 		ALIGN64 BOOL SupportEchoChannel;      /* 5184 */
 		ALIGN64 BOOL SupportDisplayControl;   /* 5185 */
 		ALIGN64 BOOL SupportGeometryTracking; /* 5186 */
-		UINT64 padding5187[1];                /* 5187 */
+		UINT64 padding5187[5188 - 5187];      /* 5187 */
 		ALIGN64 BOOL SupportVideoOptimized;   /* 5188 */
-		ALIGN64 char* RDP2TCPArgs;            /* 5189 */
+		UINT64 padding5189[5190 - 5189];      /* 5189 */
 		ALIGN64 BOOL TcpKeepAlive;            /* 5190 */
 		ALIGN64 UINT32 TcpKeepAliveRetries;   /* 5191 */
 		ALIGN64 UINT32 TcpKeepAliveDelay;     /* 5192 */
