+++ libfreerdp/common/settings_getters.c
@@ -2780,9 +2780,6 @@ const char* freerdp_settings_get_string(const rdpSettings* settings, size_t id)
 		case FreeRDP_ProxyUsername:
 			return settings->ProxyUsername;
 
-		case FreeRDP_RDP2TCPArgs:
-			return settings->RDP2TCPArgs;
-
 		case FreeRDP_ReaderName:
 			return settings->ReaderName;
 
@@ -3084,9 +3081,6 @@ char* freerdp_settings_get_string_writable(rdpSettings* settings, size_t id)
 
 		case FreeRDP_ProxyUsername:
 			return settings->ProxyUsername;
-
-		case FreeRDP_RDP2TCPArgs:
-			return settings->RDP2TCPArgs;
 
 		case FreeRDP_ReaderName:
 			return settings->ReaderName;
@@ -3398,9 +3392,6 @@ BOOL freerdp_settings_set_string_(rdpSettings* settings, size_t id, char* val, s
 
 		case FreeRDP_ProxyUsername:
 			return update_string_(&settings->ProxyUsername, cnv.c, len);
-
-		case FreeRDP_RDP2TCPArgs:
-			return update_string_(&settings->RDP2TCPArgs, cnv.c, len);
 
 		case FreeRDP_ReaderName:
 			return update_string_(&settings->ReaderName, cnv.c, len);
@@ -3730,9 +3721,6 @@ BOOL freerdp_settings_set_string_copy_(rdpSettings* settings, size_t id, const c
 
 		case FreeRDP_ProxyUsername:
 			return update_string_copy_(&settings->ProxyUsername, cnv.cc, len, cleanup);
-
-		case FreeRDP_RDP2TCPArgs:
-			return update_string_copy_(&settings->RDP2TCPArgs, cnv.cc, len, cleanup);
 
 		case FreeRDP_ReaderName:
 			return update_string_copy_(&settings->ReaderName, cnv.cc, len, cleanup);
