+++ libfreerdp/core/test/settings_property_lists.h
@@ -410,7 +410,6 @@ static const size_t string_list_indices[] = {
 	FreeRDP_ProxyHostname,
 	FreeRDP_ProxyPassword,
 	FreeRDP_ProxyUsername,
-	FreeRDP_RDP2TCPArgs,
 	FreeRDP_ReaderName,
 	FreeRDP_RedirectionAcceptedCert,
 	FreeRDP_RedirectionDomain,
